backend/easyorder/requirements.txt
Django==5.0.4
psycopg2==2.9.9
djangorestframework==3.15.1
